skills: address review — consolidate download in skills_download.py
- Move the UC skill-download client (list/fetch) and the download
  orchestration out of databricks.py and mcp.py into skills_download.py,
  so mcp.py keeps only MCP-connection glue and databricks.py only the
  shared HTTP transport helpers.
- Make --path optional: omitted, downloads write under the home dir.
- Parallelize per-skill bundle fetches; keep writes sequential (prompts).
  Add progress messages so users see skills being downloaded.
- Drop write_skill's unused return value; it prints its own outcome.
