+++ src/ucode/databricks.py
@@ -1407,123 +1407,6 @@ def build_skills_mcp_url(workspace: str, locations: list[str]) -> str:
     return base + "?" + urlencode([("schema", loc) for loc in locations])
 
 
-# --- Skill download (UC skills API + Files API) ----------------------------
-
-
-def _skill_bundle_name(skill: dict) -> str | None:
-    """The downloadable leaf name of a skill, or None if it isn't finalized.
-
-    Only finalized skills (those with a ``finalize_time``) have bundle content
-    to download. ``bundle_name`` is the leaf; fall back to the last dotted
-    segment of the resource ``name`` (``skills/<cat>.<sch>.<leaf>``).
-    """
-    if not skill.get("finalize_time"):
-        return None
-    bundle_name = skill.get("bundle_name")
-    if isinstance(bundle_name, str) and bundle_name:
-        return bundle_name
-    name = skill.get("name")
-    return name.rsplit(".", 1)[-1] if isinstance(name, str) else None
-
-
-def list_schema_skills(
-    workspace: str, token: str, catalog: str, schema: str
-) -> tuple[list[str], str | None]:
-    """List the finalized skill leaf names in ``<catalog>.<schema>``.
-
-    A non-None reason indicates the listing call itself failed.
-    """
-    hostname = workspace_hostname(workspace)
-    base_url = f"https://{hostname}/api/2.1/unity-catalog/skills"
-    query = {"parent": f"schemas/{catalog}.{schema}"}
-
-    leaves: list[str] = []
-    page_token: str | None = None
-    while True:
-        if page_token:
-            query["page_token"] = page_token
-        payload, reason = _http_get_json(f"{base_url}?{urlencode(query)}", token, timeout=30)
-        if payload is None:
-            return [], reason
-        data = cast(dict, payload) if isinstance(payload, dict) else {}
-        for skill in data.get("skills") or []:
-            leaf = _skill_bundle_name(skill) if isinstance(skill, dict) else None
-            if leaf:
-                leaves.append(leaf)
-        page_token = data.get("next_page_token")
-        if not page_token:
-            return leaves, None
-
-
-def list_skill_files(
-    workspace: str, token: str, catalog: str, schema: str, leaf: str
-) -> tuple[list[str], str | None]:
-    """List a skill bundle's files, as paths relative to the skill directory.
-
-    Recursively walks the skill's UC Volume directory (including ``SKILL.md``).
-    A non-None reason indicates the listing call itself failed.
-    """
-    hostname = workspace_hostname(workspace)
-    dirs_base = f"https://{hostname}/api/2.0/fs/directories"
-    volume_prefix = f"/Volumes/{catalog}/{schema}/{leaf}/"
-
-    relative_paths: list[str] = []
-    pending = [f"Volumes/{catalog}/{schema}/{leaf}"]
-    while pending:
-        directory = pending.pop()
-        page_token: str | None = None
-        while True:
-            url = f"{dirs_base}/{directory}"
-            if page_token:
-                url = f"{url}?{urlencode({'page_token': page_token})}"
-            payload, reason = _http_get_json(url, token, timeout=30)
-            if payload is None:
-                return [], reason
-            data = cast(dict, payload) if isinstance(payload, dict) else {}
-            for entry in data.get("contents") or []:
-                path = entry.get("path") if isinstance(entry, dict) else None
-                if not isinstance(path, str):
-                    continue
-                if entry.get("is_directory"):
-                    pending.append(path.strip("/"))
-                else:
-                    relative_paths.append(path.removeprefix(volume_prefix))
-            page_token = data.get("next_page_token")
-            if not page_token:
-                break
-    return relative_paths, None
-
-
-def fetch_skill_file(
-    workspace: str, token: str, catalog: str, schema: str, leaf: str, relative_path: str
-) -> tuple[bytes | None, str | None]:
-    """Fetch one skill bundle file's raw bytes from its UC Volume."""
-    hostname = workspace_hostname(workspace)
-    url = f"https://{hostname}/api/2.0/fs/files/Volumes/{catalog}/{schema}/{leaf}/{relative_path}"
-    return _http_get_bytes(url, token, timeout=30)
-
-
-def fetch_skill_bundle(
-    workspace: str, token: str, catalog: str, schema: str, leaf: str
-) -> tuple[dict[str, bytes] | None, str | None]:
-    """Fetch a whole skill bundle as ``{relative_path: bytes}``.
-
-    Lists the skill's files then fetches each one. All-or-nothing: a non-None
-    reason (and None bundle) means the listing or any file fetch failed, so a
-    partially-downloaded skill is never written to disk.
-    """
-    relative_paths, reason = list_skill_files(workspace, token, catalog, schema, leaf)
-    if reason:
-        return None, reason
-    bundle: dict[str, bytes] = {}
-    for relative_path in relative_paths:
-        content, reason = fetch_skill_file(workspace, token, catalog, schema, leaf, relative_path)
-        if content is None:
-            return None, reason
-        bundle[relative_path] = content
-    return bundle, None
-
-
 # Maps the gateway routing dialect a coding tool speaks to the Model Provider
 # Service `provider_type`s it can be backed by. claude speaks Anthropic's API,
 # which both the `anthropic` and `amazon_bedrock` provider types serve (Bedrock
